EvenBitesCrossoverServiceImplTest class created
